requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.8.0
Pillow==9.2.0
platformdirs==2.5.2
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
virtualenv==20.16.5
